% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computepnclu}
\alias{computepnclu}
\title{Computes the prior probability mass function for the number of clusters}
\usage{
computepnclu(Locn, Locsig, Lockappa)
}
\arguments{
\item{Locn}{Number of observation}

\item{Locsig}{Sigma parameter}

\item{Lockappa}{Kappa parameter}
}
\value{
vector of length Locn
}
\description{
Computes the prior probability mass function for the number of clusters
}
