% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computev}
\alias{computev}
\title{Computes the weights}
\usage{
computev(Locn, Locsig, Lockappa)
}
\arguments{
\item{Locn}{Number of observation}

\item{Locsig}{Sigma parameter}

\item{Lockappa}{Kappa parameter}
}
\value{
vector of length Locn
}
\description{
Computes the weights
}
